Handle exceptions in automod (#152)
Ticket: MOD-49
The automod task seems to have crashed yesterday because it didn't have
permissions in the FRU channel. As a result, it did not reschedule the
task and did not run today. I've added a block to handle any exception
so the task is still scheduled if anything happens. I've also commented
out the FRU channel for now since it won't be active till the ultimate
releases