[Dynamis.gg Import] Implement Buff component  (#133)
Ticket: NAUR-117
Issue: #103
This PR ports the Dynamis.gg buff icons + tooltips into a component we
can use.

![image](https://github.com/user-attachments/assets/0c1710ee-6c42-4a1d-bfb9-094247b6da03)

![image](https://github.com/user-attachments/assets/1582845b-deea-43d1-ae29-e30c08fdedf2)

This component defaults to the `buffs.toml` within the same directory of
the MDX file it's called in. If this needs to be overridden, you should
be able to set the relative path in `datapath`. Only `.toml` files are
supported since that's what the original supported. Supporting
`.json`/`.yaml` shouldn't be too hard, let me know if you want that
added.

Note: This does not implement shortcodes within our MDX files. Depending
on if shortcode translation is planned or not, the buff shortcodes on
dynamis.gg's content may need to be updated.
Usage (from #103):
> - b: string. The key into a (fight-specific) TOML file with buff data
in it. This is where the name, icon, game text, and explanation come
from, as well as whether the buff has the cleansable bar or not. More on
the data file in another issue.
> - dur: string. The buff duration text to render under the buff icon.
(Allowing an integer instead of a string is a helpful shorthand, but not
necessary.)
> - short: boolean (default false). If true, the name is not displayed.
> - datapath: path. Alternate path to the data file (never used, I
think).
> - stacks: optional integer. Number of stacks for the buff. If present,
the buff icon ID is offset by stacks - 1, because that's the way icon
IDs work (see [this
page](https://xivapi.com/docs/Icons?set=icons017000), for example).
> - description: optional string. Override the buff description text in
the tooltip (never used, I think).
> - explanation: optional string. Override the buff explanation text in
the tooltip (never used, I think).

Sample usage: `<Buff b="weight-of-light" dur="10s" />`