Use alternative toml package and add json option to Buffs component (#144)
Ticket: NAUR-149
Since the phases in buffs.toml could be a decimal (intermissions), it
could be an array with integers and floats. Although the newer TOML
specification allows for mixed type arrays, the older specification does
not. The TOML package we are using right now is outdated, so I've found
an alternative package to use instead that follows the most recent
specification.
I've additionally added an option to use JSON files instead