Make mobile menu subitems independent of parent (#208)
Closes #207 
This PR makes it so that items that are children of a section in the
mobile navbar (part of the accordion) are counted as separate items in
the navbar menu. This makes it so that the overall animations aren't as
jarring, e.g when clicking on the parent collapsible "Ultimate", the
animation doesn't propagate throughout the entire accordion. (Preview
can be seen in the linked issue)